优化LLM请求以减少token消耗并提高prompt cache命中
  - 移除Executor系统提示词重复拼接的全量工具说明
  - temperature/topP全链路float改double，消除0.699999988079071类精度尾巴
  - ToolRegistry三张unordered_map换std::map，tools数组顺序跨重启保持一致

+++ include/api/ApiClient.hpp
@@ -24,8 +24,8 @@ namespace LittleMeowBot::ApiClient {
     /// @return 响应文本，失败返回 std::nullopt
     drogon::Task<std::optional<std::string> > requestLLM(
         const Json::Value &messages,
-        float temperature = 1.35f,
-        float top_p = 0.92f,
+        double temperature = 1.35,
+        double top_p = 0.92,
         int max_tokens = 1024,
         const std::string &role = "memory",
         std::optional<uint64_t> sessionId = std::nullopt);
